27 jwt fix (#31)
* fix(controller): token identifier changed from username to userId

* feat(config): credentials allowed and managed by aspnet

* feat(auth): userId get from token

* fix(auth): check userId Format from token

* feat(auth): check if connected/authentified and Role

+++ CesiZen/Controllers/CategoryCommandController.cs
@@ -38,6 +38,16 @@ public class CategoryCommandController : ControllerBase
     [RoleAuthorization(Roles = "Admin")]
     public async Task<IActionResult> Create([FromBody] string name)
     {
+        if (!User.Identity?.IsAuthenticated ?? false)
+        {
+            return Unauthorized(new { message = "not authenticated" });
+        }
+
+        if (!User.IsInRole("Admin"))
+        {
+            return Forbid();
+        }
+
         var dto = new CategoryDto { Name = name };
         var result = await categoryService.Insert(dto);
 
@@ -71,6 +81,16 @@ public class CategoryCommandController : ControllerBase
     [RoleAuthorization(Roles = "Admin")]
     public async Task<IActionResult> Update(int id, [FromBody] CategoryDto dto)
     {
+        if (!User.Identity?.IsAuthenticated ?? false)
+        {
+            return Unauthorized(new { message = "not authenticated" });
+        }
+
+        if (!User.IsInRole("Admin"))
+        {
+            return Forbid();
+        }
+
         dto.Id = id;
         var result = await categoryService.Update(dto);
 
@@ -102,6 +122,16 @@ public class CategoryCommandController : ControllerBase
     [RoleAuthorization(Roles = "Admin")]
     public async Task<IActionResult> Delete(int id)
     {
+        if (!User.Identity?.IsAuthenticated ?? false)
+        {
+            return Unauthorized(new { message = "not authenticated" });
+        }
+
+        if (!User.IsInRole("Admin"))
+        {
+            return Forbid();
+        }
+
         var result = await categoryService.Delete(id);
 
         return result.Match<IActionResult>(
